+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.15.0] - 2022-08-01 - The Bare Necessities 🐻
+
 ### Added
 
 - Make "Convert For to ForEach" support Member Expressions (eg. `for (const item of foo.bar)`).
@@ -1643,7 +1645,8 @@ function doSomething(someData) {
 
 </details>
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/6.14.4...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/6.15.0...HEAD
+[6.15.0]: https://github.com/nicoespeon/abracadabra/compare/6.14.4...6.15.0
 [6.14.4]: https://github.com/nicoespeon/abracadabra/compare/6.14.3...6.14.4
 [6.14.3]: https://github.com/nicoespeon/abracadabra/compare/6.14.2...6.14.3
 [6.14.2]: https://github.com/nicoespeon/abracadabra/compare/6.14.1...6.14.2